pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager"
authors = ["Alexey-Shepelev <alexey.sailor@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
python-dotenv = "^0.21.1"
dj-database-url = "^1.2.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
django-bootstrap4 = "^22.3"
whitenoise = "^6.3.0"
django-extensions = "^3.2.1"
django-filter = "^23.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ipython = "^8.10.0"
coverage = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
